train/pyproject.toml
[project]
name = "train"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.2",
    "croniter>=6.0.0",
    "metar>=1.11.0",
    "opencv-python-headless>=4.13.0.92",
    "peft>=0.18.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "timm>=1.0.24",
    "tomli>=2.4.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
]

[project.scripts]
training = "scheduler:run_training"

[dependency-groups]
dev = [
    "tomli-w>=1.2.0",
]
